+++ PlotsBase/ext/GRExt.jl
@@ -62,7 +62,7 @@ const _gr_attrs = PlotsBase.merge_with_base_supported(
         :title,
         :window_title,
         :guide,
-        :widen,
+        :limits_modifiers,
         :lims,
         :ticks,
         :scale,